Removed ordering dependencies for EListStyleType
Removed ordering dependencies for EListStyleType from
CSSValueKeywords.h. This was already done in crrev.com/2367293002, so
this patch just updates the comments to show it is no longer needed.

Also added a comment to clarify that the fast paths still rely on the
ordering, so not to update those.

BUG=665272

Review-Url: https://codereview.chromium.org/2535663002
Cr-Original-Commit-Position: refs/heads/master@{#435192}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 02bf8ffd9e09a701ca30410df5e1b71d37b0dd09