Worker: Fix possible UAF in closures posted from ThreadedWorkletObjectProxy
This fix is the same with the following CL for InProcessWorkerObjectProxy
created by alexclarke@chromium.org: https://codereview.chromium.org/2478113002

Tasks posted by ThreadedWorkletObjectProxy can be arbitrarily delayed so it's
not safe to post a closure with an crossThreadUnretained pointer to the
messaging proxy because we can't guarantee it's still there when the closure
eventually gets run.

We fix this with a WeakPtr.

BUG=660427

Review-Url: https://codereview.chromium.org/2534803003
Cr-Original-Commit-Position: refs/heads/master@{#435208}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 7f239c0f458773e19039218f566590a44aeb0f87