Raster: Allow a RasterSource to specify its color space
The pictures recorded by GraphicsLayers in Blink have a particular
implied color space in the recorded content. This is the color space
that the decoded images are in (and soon to be the color space that
CSS colors are converted into from sRGB).

We need to communicate that information from Blink to cc, so that
the resulting cc::Resources can be tagged with the appropriate color
space.

Plumb this through the same channel in WebDisplayItemList that is
used for "IsSuitableForGpuRasterization".

Of note is that "true color mode" content (content that does linear
blending) has the luxury of being able to decide the target color
space at raster time -- it does not have an implied color space.

BUG=667411
CQ_INCLUDE_TRYBOTS=master.tryserver.blink:linux_trusty_blink_rel

Review-Url: https://codereview.chromium.org/2525453002
Cr-Original-Commit-Position: refs/heads/master@{#435212}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b7210d35c666575a48ed600d7758c451220b09b8