move blink::StyleBuilderConverter::convertPositionLength into header file
blink::StyleBuilderConverter::convertPositionLength is a function
template that is used from outside StyleBuilderConverter.cpp. Currently,
that happens to work, because the only invocations from outside the file
happen to use the same template parameters as the ones inside, and the
calls from inside the file are not being optimized away. However, this
may not continue to be true, so this change moves the template into the
header file so that it's always available when needed.

Review-Url: https://codereview.chromium.org/2533273005
Cr-Original-Commit-Position: refs/heads/master@{#435452}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 39d7d7cde2cdfa882746f744b489db27be8aae7e