Fix memory corruption related to load blocking resource move
This fixes an issue where a resource loader belonging to one
ResourceFetcher was accidentally added as a blocking loader to another
ResourceFetcher, by checking the loader is part of the already
non-blocking loaders belonging to current ResourceFetcher.

This also adds DCHECKs on a couple of methods removing loaders from
hashmaps, to make sure we're not trying to remove a nullptr.

BUG=666563

Review-Url: https://codereview.chromium.org/2537303003
Cr-Original-Commit-Position: refs/heads/master@{#435573}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 7497990e6eb19dc8dd61de4f188553c9c054cef9