Fix a minor issue in OffscreenCanvas.drawImage
This CL fixes the problem in OffscreenCanvas.drawImage function when it
takes another OffscreenCanvas without a context as its image source. In
this case, it should draw transparent black. However, currently this is
not currently handled.

This CL also adds a layout test to verify the case.

BUG=670393

Review-Url: https://codereview.chromium.org/2550443002
Cr-Original-Commit-Position: refs/heads/master@{#435723}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: c5164dcf766dffdbbe6575e7c7c68cb920698c8a