Screen Orientation: add ScreenOrientationController in core/ with Impl in modules/.
It allows code able to access core/ to manipulate screen orientation
instead of it being limited to modules/ and web/. It is not part of
platform/ because it requires LocalFrame access.

This will be used by MediaControls to lock/unlock the screen orientation.

BUG=326572

Review-Url: https://codereview.chromium.org/1911163003
Cr-Original-Commit-Position: refs/heads/master@{#435760}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 36c92fada7aa67ff35ff8afd4473bcbed2e336ca