[Devtools] Rename logEntries to node in network
This patch is in prepopration for groupping in network panel. Before we
introduced logEntries as an interface, but this is proving to difficult
to work with and complicating the code. We decided to move to use
NetworkDataGridNode instead of an interface. This patch is just a
rename.

R=dgozman
BUG=666971

Review-Url: https://codereview.chromium.org/2548663002
Cr-Original-Commit-Position: refs/heads/master@{#435831}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 2a92e8ab145d34cea6dfc94f9084e78e2724512a