Fix WebBluetooth generator to distinguish between different methods.
This change allows tests to be created that exercise both
characteristic.readValue/writeValue and descriptor.readValue/writeValue.
script-tests's name determine what feature they are testing.  For example, a
test that wants to tests gatt operations on a descriptor can be named:

bluetooth/script-tests/descriptor/<>

then in bluetooth/descriptor/<>/ each of the gatt calls will be generated.

In addition and prep for this change, we are moving files around.  The
structure now is the following:

device
    gattserverdisconnected-event

service
   getCharacteristic
   getCharacteristics

server
    getPrimaryService
    getPrimaryServices
    connect
    disconnect

characterisitic
   startNotifications
   stopNotifications
   readValue
   writeValue
   getDescriptor
   getDescriptors
   notifications

BUG=668836
R=ortuno

Review-Url: https://codereview.chromium.org/2544513004
Cr-Original-Commit-Position: refs/heads/master@{#436173}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 36c5c1c79009ee041372a8487ec3317606e9f5f4