Changed Order to an enum class and renamed its members
Changed Order to an enum class and gave it an unsigned underlying type.
Also renamed it from Order to EOrder to match the other ComputedStyle
enums, and renamed its members to match its keywords from
CSSValueKeywords.in. Also removed the '= 0' value setter for Logical and
the one callsite where this was used (implicit int conversion is not
allowed for enum classes anyway).

Changing it to an enum class enforces better namespacing and code
practices. Adding the unsigned underlying type is pre-work for when the
class is eventually stored as an enum bitfield (it would be done in this
patch, except a presubmit warning already exists that prevents that. The
presubmit warning needs to be updated before that change can occur.)

This is also pre-work to move EOrder to be generated in
ComputedStyleBase.

BUG=628043

Review-Url: https://codereview.chromium.org/2546333002
Cr-Original-Commit-Position: refs/heads/master@{#436233}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 5fcf8dbc81c32665ab015b7f0cf303f3427a3d93