Inline LayoutRect::enclosingIntRect
Noticed this function was hot on blink_perf.paint's color-changes.html
microbenchmark. With this patch, color-changes.html runs 2.7% faster
(average -0.8ms) but that number should be taken with a grain of salt
since it's near the standard deviation of the test.

Manual profiling with and without the patch confirmed this function is
no longer hot.

Review-Url: https://codereview.chromium.org/2550083002
Cr-Original-Commit-Position: refs/heads/master@{#436323}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: af2c8d1e6a537c356b4cf931bb5aca1e5f09f119