[MediaElement] Allow preloading for non-network scheme src
This CL weakens the restriction of media element preloading. The
reason for this change is that the preload attribute for media
elements is overriden as "none" when data saver is
enabled. However there is no need to apply this override for
non-network media sources, since they do not cost network
traffic.

BUG=661220

Review-Url: https://codereview.chromium.org/2546993003
Cr-Original-Commit-Position: refs/heads/master@{#436358}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f9d8e609f1298c6a563274fbdbe6ee5c45df3034