Pull up alpha correction in GIF decoder to ImageDecoder.
Just a refactor. No behavioral changes, no added tests.

Pull up the code section in GIFImageDecoder::frameComplete, which
corrects the alpha setting of the frame buffer if the entire frame
is opaque, to ImageDecoder::correctAlphaWhenFrameBufferSawNoAlpha.
This makes the code reusable for the (A)PNG decoder.

Review-Url: https://codereview.chromium.org/2551543002
Cr-Original-Commit-Position: refs/heads/master@{#436363}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 5ff7b531728f9e39581b1a6b57be4d432711cb40