WebKit: remove DCHECK_IS_ON from FileWriterSync.
This CL removes checks for DCHECK_IS_ON in FileWriterSync as they are not used.
(DCHECK macro requires its arguments to be buildable even if check itself is
disabled.) So this CL fixes build in case we want debug build with DCHECKs
disabled.

R=nhiroki@chromium.org

BUG=666660

Review-Url: https://codereview.chromium.org/2524813003
Cr-Original-Commit-Position: refs/heads/master@{#436473}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: c9600cd059f7aa0231a89fa2817c4ffca0454122